feat: add decimal conversion for cross-chain route amounts
Source and destination tokens may have different decimals (e.g., USDC
on Solana = 6, USDC on BSC = 18). After fee calculation, route_amount
is now converted from source to destination decimals. Uses u128 to
handle scaling up without overflow (e.g., 6→18 dec multiplies by 10^12).

+++ svm/programs/swap-intent/src/events.rs
@@ -9,7 +9,7 @@ pub struct IntentCreated {
     pub user: Pubkey,
     pub reward_token: Pubkey,
     pub swap_output: u64,
-    pub route_amount: u64,
+    pub route_amount: u128,
     pub destination: u64,
 }
